utils/auth.js created
The purpose of this code is to provide an authentication system for a Node.js application using JSON Web Tokens. The authMiddleware function is used to verify the presence and validity of the JWT in incoming requests, and the signToken function is used to generate a new JWT with the user's data.