CRUD operation created/server-side API
The purpose of this code is to define the API routes for the exercise tracking application. The routes are separated into two main categories: Cardio and Resistance exercises. Each route is associated with a specific controller function that handles the corresponding CRUD (Create, Read, Update, Delete) operations and The authMiddleware is applied to all routes.